Put all tests for OpenMetrics V2 in the same file
In preparation for adding tests for dynamic parsing based on content type
